Reducing Cache Stampeding on Restart
When the cache is empty, the call to `fetch()` will evaluate the passed
getter function which is then cached. Until the getter function finishes
executing, the cache will remain empty.

If multiple requests are made in a short period of time, then it's possible
for the getter function to be called multiple times. This is especially
problematic if the getter function is expensive or takes a long time to
execute.

To limit the "thundering herd" scenario, the getter function is moved from
the `fetch()` module function into the `handle_call` GenServer callback.

Since the `handle_call` callbacks are blocking, any additional calls to the
cache that are received while the getter function is being executed will be
queued until after it completes.

The `handle_call` callback then stores the result in the cache. Many of the
queued calls will be able to read the value from the cache instead of executing
the getter function again.